:gear: Reduce goldilock default build by introducing BUILD_TESTING and making tests builds optional.
When HFC provisions goldilock on a platform where pre-released binaries are
incompatible, it builds with default CMake flags, which did build all test
with it which is wasteful in such scenario.

CMake common practice is to rely on BUILD_TESTING, which would be
automatically defined  when doing `include(CTest)`, this would
however set a BUILD_TESTING option with the default to ON
 - c.f. https://cmake.org/cmake/help/latest/module/CTest.html

We think that especially for goldilock which is more like a dependency
than a main project being worked on for it's main target audience we
prefer BUILD_TESTING being OFF by default.